==> 0_inspect_data.ipynb <==
### Inspect data
#### List of tasks accomplished in this Jupyter Notebook:
- Output number of live and dead animals for each species and treatment
- Remove dead larvae from experiment master file
- Check that manually annotated files all exist and are spelled correctly

import numpy as np
import pandas as pd
import glob, os

# READ IN CLEANED DATASET and PRINT NUMBER OF ANIMALS FOR EACH TREATMENT
df = pd.read_csv("./data/experiment_IDs/static_data.csv")
experiments = df["treatment_odor"].unique()
species = df['species'].unique()
for specie in species: 
    temp = df[df['species'] == specie]
    temp = temp[temp['dead']=='no']
    print('Total number of animals:', len(temp))
    for experiment in experiments:
        temp = df[df["treatment_odor"] == experiment]
        temp = temp[temp['species'] == specie]
        temp2 = temp[temp['dead']!='yes']
        temp3 = temp[temp['dead']=='yes']
        print('Live n=', str(len(temp2)), 'dead n=', str(len(temp3)),
              ': total n=', str(len(temp)), 'unknown n=', 
              str(len(temp)-len(temp3)-len(temp2)),
              'for', experiment, specie)
    print('---')

# REMOVE DEAD LARVAE AND SAVE DATA IN NEW SPREADSHEET
df = pd.read_csv('./data/experiment_IDs/static_data.csv')
print(len(df), "total larvae experiments")

no_dead = df[df['dead'] != 'yes'].copy()
print(len(no_dead), "larvae after removing dead larvae")

no_dead.to_csv("./data/experiment_IDs/cleaned_static_data.csv", index=None)
print("--- Data cleaned and saved to file ---")
display(no_dead.tail())

# CHECK THAT MANUALLY ANNOTATED FILES ALL EXIST AND ARE SPELLED CORRECTLY
df = pd.read_csv("./data/trajectories/manually_checked_beginning_pause.csv")

for name in df["filename"].values:
    fname = "./data/trajectories/video_csvs/"+name+".csv"
    if not os.path.isfile(fname):
        print(fname)
        
print("--- All checks passed ---")
