==> 4_calculate_experiment_values_per_larvae.ipynb <==
### Calculate Experiment Analyses Values Per Larvae
#### List of tasks accomplished in this Jupyter Notebook:
- Calculate the median concentration preferred by the larvae (normalized to the acclimation period)
- Calculate the discovery time for each larvae (normalized to the acclimation period)
- Calculate the speed in areas of high concentration, compared to speed in areas of low concentration, normalized to behavior during the acclimation period. 
- Output all of these values into a separate spreadsheet. 

import numpy as np
import pandas as pd
import math, os
import more_itertools as mit

concentration_cutoff = 50

fps = 2
sh = 1 # direction in which to shift vector for delta calculations
threshold = 2 # consecutive seconds spent moving
still_threshold = 10 # seconds for minimum still num count
angle_thresh = 20 # Allowed error from previous angle
sharp_angle = 45
spiral_thresh = 4
spiral_speed_thresh = np.inf

def safe_divide(x, y):
    x = safe_get(x)
    y = safe_get(y)
    try: 
        ans = x / y
    except ZeroDivisionError: 
        return 0
    return ans
    
def safe_get(x, ans=0):
    if math.isnan(x):
        return ans
    return x

df = pd.read_csv("./data/experiment_IDs/cleaned_static_data.csv")

animals = df["animal_ID"].unique()
# Check that animal IDs in dataframe are unique
assert len(df) == len(animals)

fps = 2
a_master_df = pd.DataFrame()
e_master_df = pd.DataFrame()

for index, row in df.iterrows():
    animal = row['animal_ID']
    aID = animal[:9]
    pos = animal[10:]
    
    exp_filename = "./data/trajectories/video_calculations/"+aID+'-E-'+pos+'.csv'
    acc_filename = "./data/trajectories/video_calculations/"+aID+'-A-'+pos+'.csv'
    
    for tag, filename, master_df in zip(['A_', 'E_'], 
                             [acc_filename, exp_filename], 
                             [a_master_df, e_master_df]):
        if os.path.isfile(filename):
            temp = pd.read_csv(filename)

            # MEDIAN CONCENTRATION -----------------------------
            median_conc = safe_get(temp["concentration"].median())

            # DISCOVERY TIME IN MINUTES ------------------------
            if len(temp[temp["concentration"] >= concentration_cutoff]) > 0:
                discovery_time = safe_divide(temp[temp["concentration"] >= concentration_cutoff]["frames"].values[0], fps*60)
            else:
                discovery_time = 15

            # Divide up into thresholds for concentration and concentration differences
            high = temp[temp["concentration"] >= 50]
            low = temp[temp["concentration"] < 50]
            high_move = high[high["moving"] == True]
            low_move = low[low["moving"] == True]
            
            # SPEED IN BODY LENGTHS ----------------------------
            # concentration dependent
            high_speed = safe_divide(high_move["speed_BL"].median(), row["larvae_length_mm"])
            low_speed = safe_divide(low_move["speed_BL"].median(), row["larvae_length_mm"])
            c_speed = high_speed - low_speed
 
            temp = pd.DataFrame({# STATIC VARIABLES --------------------------
                                "animal_ID": [row["animal_ID"]], 
                                "treatment_odor": row["treatment_odor"],
                                "sex": row["sex"], 
                                "species": row['species'],
                                'dead':row['dead'],

                                # VARIABLES CALCULATED FOR ENTIRE TIME -------
                                tag+"median_conc": median_conc, 
                                tag+"discovery_time": discovery_time,
                                tag+"c_speed": c_speed
                                 })
            if tag == 'A_':
                a_master_df = pd.concat([a_master_df, temp])
            elif tag == 'E_':
                e_master_df = pd.concat([e_master_df, temp])

display(a_master_df.head(2))
display(e_master_df.head(2))
a_master_df.drop(['treatment_odor', 'sex', 'species', 'dead'], axis=1, inplace=True)
df = pd.merge(a_master_df, e_master_df, on='animal_ID', how='inner')

df["median_conc_diff"] = df["E_median_conc"] - df["A_median_conc"]
df["discovery_time_diff"] = df["E_discovery_time"] - df["A_discovery_time"] 
df["c_speed_diff"] = df["E_c_speed"] - df["A_c_speed"]

df.to_csv('./data/trajectories/cleaned_animal_analyses_experiment.csv', index=False)
display(df.tail())
