==> 6_statistical_analyses.ipynb <==
### Statistical Analyses
#### List of tasks accomplished in this Jupyter Notebook:
- Calculate if there are any differences among the six species for each of the ten navigational variables (Kruskal test)
- Calculate if the larvae of each species prefer each of the different treatment odors
- Check if data gathered has a normal distribution (to inform the types of statistical tests that can be conducted)
- Calculate if there are mortality differences across the larval species (Fisher's exact test for binary data)
- Calculate pairwise perMANOVA for each species against each other species (navigation variables) as well as pairwise tests of multivariate dispersion (ANOVA) for all of the ten navigation variables during the acclimation period. 

R.version # print R version for reference
library(vegan)

setwd("/Users/ghost/Documents/Gitrepos/aedes-aegypti-2020/data/")
method_n <- "holm"

species <- c("Aedes aegypti", "Aedes albopictus", "Anopheles arabiensis", 
             "Anopheles gambiae", "Culex quinquefasciatus", "Culex tarsalis")

# FIGURE S2 and FIGURE 2

# Adjust all tests for the same data using the Holm-Bonferroni correction
p <- "placeholder"
t <- "placeholder"

data <- read.csv('./trajectories/cleaned_animal_analyses_acclimation.csv')
data <- subset(data, data$dead!='yes')
data$species <- as.factor(data$species)

# Individual variables by species
resp <- kruskal.test(data$time_move_p~data$species)
p <- c(p, resp$p.value)
t <- c(t, "time_move_p by species")

resp <- kruskal.test(data$total_dist_m~data$species)
p <- c(p, resp$p.value)
t <- c(t, "total_dist_m by species")

resp <- kruskal.test(data$avg_speed_BL~data$species)
p <- c(p, resp$p.value)
t <- c(t, "avg_speed_BL by species")

resp <- kruskal.test(data$max_speed_BL~data$species)
p <- c(p, resp$p.value)
t <- c(t, "max_speed_BL by species")

resp <- kruskal.test(data$mean_speed_first_BL~data$species)
p <- c(p, resp$p.value)
t <- c(t, "mean_speed_first_BL by species")

resp <- kruskal.test(data$diff_speed_first_last_BL~data$species)
p <- c(p, resp$p.value)
t <- c(t, "diff_speed_first_last_BL by species")

resp <- kruskal.test(data$sharp_turns_p~data$species)
p <- c(p, resp$p.value)
t <- c(t, "sharp_turns_p by species")

resp <- kruskal.test(data$spirals~data$species)
p <- c(p, resp$p.value)
t <- c(t, "spirals by species")

resp <- kruskal.test(data$continuous~data$species)
p <- c(p, resp$p.value)
t <- c(t, "continuous by species")

resp <- kruskal.test(data$max_still_sec~data$species)
p <- c(p, resp$p.value)
t <- c(t, "max_still_sec by species")

resp <- kruskal.test(data$length_mm~data$species)
p <- c(p, resp$p.value)
t <- c(t, "length_mm by species")

"TRUE/FALSE: Larval length data has a normal distribution"
"All tests must pass to use statistical tests that assume normality"
for (specie in species){
    spec <- subset(data, species==specie)
    print(shapiro.test(spec$length_mm)$p.value>0.05)}

"TRUE/FALSE: Larval length data has a normal distribution"
"All tests must pass to use statistical tests that assume normality"
for (specie in species){
    spec <- subset(data, species==specie)
    print(shapiro.test(spec$spirals)$p.value>0.05)}

# remove placeholder values and print out p values
p <- p[-1]; 
t <- t[-1];
cat("Number of tests ran:", length(p))
p <- p.adjust(p, method=method_n)
d <- data.frame(t, p, p<0.05, p<0.01, p<0.001, p<0.0001)
d

# FIGURE 4A

data <- read.csv('./trajectories/cleaned_animal_analyses_experiment.csv')
data <- subset(data, data$dead!='yes')
treatments <- c("05_percent_food", "100ul_quinine", '100ul_milliQ_water')

data$treatment_odor <- as.factor(data$treatment_odor)
data$sex <- as.factor(data$sex)
data$species <- as.factor(data$species)

for (specie in species){
    spec <- subset(data, species==specie)
    print(specie)
    ps <- 'test'
    ts <- 'test'
    ds <- 'test'
    for (treat in treatments){
        ss <- subset(spec, treatment_odor==treat)
        resp <- t.test(ss$A_median_conc, ss$E_median_conc, paired=TRUE, 
                        alternative="two.sided")
        diff <- mean(ss$E_median_conc) - mean(ss$A_median_conc)
        ps <- c(ps, resp$p.value)
        ts <- c(ts, treat)
        ds <- c(ds, diff)}
    # remove placeholder values and print out p values
    ps <- ps[-1]; 
    ts <- ts[-1];
    ds <- ds[-1];
    d <- data.frame(ts, ds, ps)
    print(d)}

# FIGURE 4B

data <- read.csv('./trajectories/cleaned_animal_analyses_experiment.csv')
data <- subset(data, data$dead!='yes')
treatments <- c("05_percent_food", "100ul_quinine", "100ul_milliQ_water")

data$treatment_odor <- as.factor(data$treatment_odor)
data$sex <- as.factor(data$sex)
data$species <- as.factor(data$species)

"TRUE/FALSE: Data queried has a normal distribution"
"All tests must pass to use statistical tests that assume normality"

for (specie in species){
    spec <- subset(data, species==specie)
    for (treat in treatments){
        spec_t <- subset(spec, treatment_odor==treat)
        print(shapiro.test(spec_t$E_discovery_time)$p.value>0.05)}}

for (specie in species){
    spec <- subset(data, species==specie)
    water <- subset(spec, treatment_odor=='100ul_milliQ_water')
    resp <- kruskal.test(spec$discovery_time_diff~spec$treatment_odor)
    
    print("----")
    print(specie)
    print(resp$p.value)
}

# FIGURE S2

# Adjust all tests for the same data using the Holm-Bonferroni correction
p <- "placeholder"
t <- "placeholder"

mortality <- read.csv('./experiment_IDs/species_mortality_table.csv',
                      header=TRUE, row.names=1)
mortality

# Fisher's exact test for mortality data, between species
resp <- fisher.test(mortality, simulate.p.value=TRUE)
p <- c(p, resp$p.value)
t <- c(t, paste("Species differences", sep=" "))

# Fisher's exact test for mortality data, between treatments for each species
data <- read.csv('./experiment_IDs/experiment_mortality_table.csv', header=TRUE)
for (specie in species){
    data_subset <- subset(data, species==specie, select=c(alive, dead))
    resp <- fisher.test(data_subset, simulate.p.value=TRUE)
    p <- c(p, resp$p.value)
    t <- c(t, paste(specie, sep=" "))}

# remove placeholder values and print out p values
p <- p[-1]; 
t <- t[-1];

# cat("Number of tests ran:", length(p))
p <- p.adjust(p, method=method_n)
d <- data.frame(t, p, p<0.05, p<0.01, p<0.001, p<0.0001)
print(d)

# FIGURE 3
n <- 10000

# perMANOVA for all six species at once
data <- read.csv('./trajectories/cleaned_animal_analyses_acclimation.csv')
sp.data <- subset(data, select=-c(species, treatment_odor, sex, dead, length_mm))
sp.species <- subset(data, select=c(animal_ID, species))
rownames(sp.data) <- sp.data$animal_ID
sp.data <- subset(sp.data, select=-c(animal_ID))
rownames(sp.species) <- sp.species$animal_ID
sp.species <- subset(sp.species, select=-c(animal_ID))
sp.data.tra<-decostand(sp.data, method='standardize', margin='column', plot=F)
perm <- adonis(sp.data.tra~species, data=sp.species, 
               permutations=n, method='euclidean')
# perm[[1]]
print("----")
print("Comparison across all species")
perm
print("----")

# Adjust all tests for the same data using the Holm-Bonferroni correction
p <- "placeholder"
t <- "placeholder"
f <- "placeholder"

sp_pairs <- list()
sp_pairs[[1]] <- c('Aedes aegypti', 'Aedes albopictus')
sp_pairs[[2]] <- c('Aedes aegypti', 'Anopheles arabiensis')
sp_pairs[[3]] <- c('Aedes aegypti', 'Anopheles gambiae')
sp_pairs[[4]] <- c('Aedes aegypti', 'Culex tarsalis')
sp_pairs[[5]] <- c('Aedes aegypti', 'Culex quinquefasciatus')
sp_pairs[[6]] <- c('Aedes albopictus', 'Anopheles arabiensis')
sp_pairs[[7]] <- c('Aedes albopictus', 'Anopheles gambiae')
sp_pairs[[8]] <- c('Aedes albopictus', 'Culex tarsalis')
sp_pairs[[9]] <- c('Aedes albopictus', 'Culex quinquefasciatus')
sp_pairs[[10]] <- c('Anopheles arabiensis', 'Anopheles gambiae')
sp_pairs[[11]] <- c('Anopheles arabiensis', 'Culex tarsalis')
sp_pairs[[12]] <- c('Anopheles arabiensis', 'Culex quinquefasciatus')
sp_pairs[[13]] <- c('Anopheles gambiae', 'Culex tarsalis')
sp_pairs[[14]] <- c('Anopheles gambiae', 'Culex quinquefasciatus')
sp_pairs[[15]] <- c('Culex tarsalis', 'Culex quinquefasciatus')

for (pair in sp_pairs){
    data <- read.csv('./trajectories/cleaned_animal_analyses_acclimation.csv')
    data <- subset(data, species == pair[1] | species == pair[2])
    data$species <- as.factor(data$species)
    sp.data <- subset(data, select=-c(species, treatment_odor, sex, dead, length_mm))
    sp.species <- subset(data, select=c(animal_ID, species))
    rownames(sp.data) <- sp.data$animal_ID
    sp.data <- subset(sp.data, select=-c(animal_ID))
    rownames(sp.species) <- sp.species$animal_ID
    sp.species <- subset(sp.species, select=-c(animal_ID))
    sp.data.tra<-decostand(sp.data, method='standardize', margin='column', plot=F)
    perm <- adonis(sp.data.tra~species, data=sp.species, 
                   permutations=n, method='euclidean')
    p <- c(p, perm[[1]][1,6])
    f <- c(f, perm[[1]][1,4])
    t <- c(t, paste(c('PM', pair), collapse=', '))
}

for (pair in sp_pairs){
    data <- read.csv('./trajectories/cleaned_animal_analyses_acclimation.csv')
    data <- subset(data, species == pair[1] | species == pair[2])
    data$species <- as.factor(data$species)
    sp.data <- subset(data, select=-c(species, treatment_odor, sex, dead, length_mm))
    sp.species <- subset(data, select=c(animal_ID, species))
    rownames(sp.data) <- sp.data$animal_ID
    sp.data <- subset(sp.data, select=-c(animal_ID))
    rownames(sp.species) <- sp.species$animal_ID
    sp.species <- subset(sp.species, select=-c(animal_ID))
    sp.data.tra<-decostand(sp.data, method='standardize', margin='column', plot=F)
    sp.dist <- vegdist(sp.data.tra,"euclidean")
    sp.bdp <- betadisper(sp.dist, sp.species$species)
    an <- anova(sp.bdp)
    p <- c(p, an[1, 5])
    f <- c(f, an[1, 4])
    t <- c(t, paste(c('BDP', pair), collapse=', '))
}

# remove placeholder values and print out p values
p <- p[-1]; 
t <- t[-1];
f <- f[-1];

# cat("Number of tests ran:", length(p))
p <- p.adjust(p, method=method_n)
d <- data.frame(t, p, f)
print(d)

# Test pairwise code for just one pair
pair <- c('Anopheles gambiae', 'Culex quinquefasciatus')

data <- read.csv('./trajectories/cleaned_animal_analyses_acclimation.csv')
data <- subset(data, species == pair[1] | species == pair[2])
data$species <- as.factor(data$species)
sp.data <- subset(data, select=-c(species, treatment_odor, sex, dead, length_mm))
sp.species <- subset(data, select=c(animal_ID, species))
rownames(sp.data) <- sp.data$animal_ID
sp.data <- subset(sp.data, select=-c(animal_ID))
rownames(sp.species) <- sp.species$animal_ID
sp.species <- subset(sp.species, select=-c(animal_ID))
sp.data.tra<-decostand(sp.data, method='standardize', margin='column', plot=F)
perm <- adonis(sp.data.tra~species, data=sp.species, 
               permutations=100, method='euclidean')
PM_p <- perm[[1]][1,6]
PM_f <- perm[[1]][1,4]

sp.dist <- vegdist(sp.data.tra,"euclidean")
sp.bdp <- betadisper(sp.dist, sp.species$species)
an <- anova(sp.bdp)

BDP_p <- an[1, 5]
BDP_f <- an[1, 4]

# Print everything 
perm[[1]]
PM_p
PM_f
an
BDP_p
BDP_f
